refactor(git): delete git_operations.py shim, update test imports
Remove the backward-compat re-export shim. Test imports now point
directly at git/ submodules (git.operations, git.validation,
git.patches, git.clone, git.archive).
